updating default settings and tests
