+++ app/scripts/components/common/page-footer/footer.test.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import React, { ComponentType } from 'react';
 import { render, screen } from '@testing-library/react';
-import { navItems, footerSettings } from '../../../../../mock/veda.config.js';
+import { navItems } from '../../../../../mock/veda.config.js';
 import NasaLogoColor from '../nasa-logo-color';
 import { NavItem } from '../page-header/types.js';
 
@@ -8,31 +8,56 @@ import PageFooter from './index';
 
 const mockMainNavItems: NavItem[] = navItems.mainNavItems;
 const mockSubNavItems: NavItem[] = navItems.subNavItems;
-const mockFooterSettings = footerSettings;
+// const mockFooterSettings = footerSettings;
 const hideFooter = false;
+const mockLinkProperties = {
+  pathAttributeKeyName: 'to',
+  LinkElement: 'a' as unknown as ComponentType
+};
+jest.mock('./default-config', () => ({
+  footerSettings: {
+    secondarySection: {
+      division: 'NASA EarthData 2024',
+      version: 'BETA VERSION',
+      title: 'NASA Official',
+      name: 'test',
+      to: 'test@example.com',
+      type: 'email'
+    },
+    returnToTop: true
+  }
+}));
 
 describe('PageFooter', () => {
-  beforeEach(() => {
-    render(
-      <PageFooter
-        settings={mockFooterSettings}
-        mainNavItems={mockMainNavItems}
-        subNavItems={mockSubNavItems}
-        hideFooter={hideFooter}
-        logoSvg={<NasaLogoColor />}
-      />
-    );
-  });
   afterEach(() => {
     jest.clearAllMocks();
   });
   test('renders the PageFooter', () => {
+    render(
+      <PageFooter
+        mainNavItems={mockMainNavItems}
+        subNavItems={mockSubNavItems}
+        hideFooter={hideFooter}
+        logoSvg={<NasaLogoColor />}
+        linkProperties={mockLinkProperties}
+      />
+    );
     const footerElement = document.querySelector('footer');
 
     expect(footerElement).toBeInTheDocument();
     expect(footerElement).not.toHaveClass('display-none');
   });
+
   test('renders correct buttons and links', () => {
+    render(
+      <PageFooter
+        mainNavItems={mockMainNavItems}
+        subNavItems={mockSubNavItems}
+        hideFooter={hideFooter}
+        logoSvg={<NasaLogoColor />}
+        linkProperties={mockLinkProperties}
+      />
+    );
     expect(screen.getByText('Data Catalog')).toBeInTheDocument();
     expect(screen.getByText('Exploration')).toBeInTheDocument();
     expect(screen.getByText('Stories')).toBeInTheDocument();
@@ -43,26 +68,23 @@ describe('PageFooter', () => {
 });
 
 describe('PageFooter dynamic settings', () => {
-  test('Return to top does not show', () => {
-    const updatedMockSettings = { ...mockFooterSettings, returnToTop: false };
-
-    render(
-      <PageFooter
-        settings={updatedMockSettings}
-        mainNavItems={mockMainNavItems}
-        subNavItems={mockSubNavItems}
-        hideFooter={hideFooter}
-        logoSvg={<NasaLogoColor />}
-      />
-    );
-    expect(() => screen.getByText('Return to top')).toThrow();
-  });
   test('Hide footer should function correctly', () => {
-    const updatedMockSettings = { ...mockFooterSettings, returnToTop: false };
-
+    jest.mock('./default-config', () => ({
+      footerSettings: {
+        secondarySection: {
+          division: 'NASA EarthData 2024',
+          version: 'BETA VERSION',
+          title: 'NASA Official',
+          name: 'test',
+          to: 'test@example.com',
+          type: 'email'
+        },
+        returnToTop: true
+      }
+    }));
     render(
       <PageFooter
-        settings={updatedMockSettings}
+        linkProperties={mockLinkProperties}
         mainNavItems={mockMainNavItems}
         subNavItems={mockSubNavItems}
         hideFooter={true}
